feat: delete the course-shell areas from the platform — they ship as Edu-plugin source now
Phase 3 of the education extraction. EducationLayoutAreas (the four
course-shell areas: StartExercise / GoToMyCopy / CourseNav / Learn) and
its test are DELETED from the platform: the areas ship as in-mesh Edu-
plugin source (CourseShellAreas, Plugins#481, Edu v1.0.29, merged and
rolling out via plugin auto-update), registered on the Edu NodeTypes —
course roots are Store/Plugin-typed and use the Store cover, so nothing
platform-side renders these areas any more.

What remains compiled in MeshWeaver.Education: the navigation
contributor (EducationNavigationProvider — mesh-level DI, the documented
plugin ceiling), CourseProgress, TrainingSimResponder (paired with
PromptCell.Live, whose in-mesh callers pin it until the PromptCell
migration), and the seven shared query/path helpers the contributor used
from the deleted class, extracted into EducationHelpers.

Verified: MeshWeaver.Education + Memex.Portal.Shared build -c Release
-warnaserror clean; AI.Test and Graph.Test build;
EducationNavigationProvider tests 11/11.
